pyproject.toml
[project]
name = "my-notebook"
version = "0.1.0"
description = "A simplified AI-powered research assistant"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.9.2",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "surrealdb>=1.0.4",
    "langchain>=1.2.0",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "langchain-openai>=1.1.6",
    "langchain-google-genai>=4.1.2",
    "esperanto>=2.19.7,<3",
    "content-core>=1.14.1,<2",
    "ai-prompter>=0.3,<1",
    "podcast-creator>=0.12.0,<1",
    "tiktoken>=0.12.0",
    "httpx[socks]>=0.27.0",
    "numpy>=2.4.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"core" = "core"}

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "E402", "E722"]
